- less strenuous water animation - added size key to bobs - building capabilities now honour nearby objects
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@219 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/game.h
@@ -78,7 +78,7 @@ public:
 
    inline Map *get_map() { return map; }
 
-   inline Player* get_player(ushort n) { assert(n>=1 && n<=MAX_PLAYERS); return m_players[n-1]; }
+   inline Player* get_player(int n) { assert(n>=1 && n<=MAX_PLAYERS); return m_players[n-1]; }
 	bool get_allow_cheats();
 	
 public:
